+++ src/circular_buffer.c
@@ -2,7 +2,7 @@
 
 static inline bool cb_is_power_of_two(size_t value) { return value != 0 && (value & (value - 1)) == 0; }
 
-void cb_init(circular_buffer_t* cb, uint8_t* storage, size_t capacity) {
+void cb_init(circular_buffer_t* cb, uint32_t* storage, size_t capacity) {
   cb->data = storage;
   cb->capacity = capacity;
   cb->head = 0;
@@ -25,7 +25,7 @@ size_t cb_size(const circular_buffer_t* cb) { return cb->size; }
 
 size_t cb_capacity(const circular_buffer_t* cb) { return cb->capacity; }
 
-bool cb_push(circular_buffer_t* cb, uint8_t value) {
+bool cb_push(circular_buffer_t* cb, uint32_t value) {
   if (cb->size == cb->capacity || cb->capacity == 0) {
     return false;
   }
@@ -40,7 +40,7 @@ bool cb_push(circular_buffer_t* cb, uint8_t value) {
   return true;
 }
 
-bool cb_pop(circular_buffer_t* cb, uint8_t* value) {
+bool cb_pop(circular_buffer_t* cb, uint32_t* value) {
   if (cb->size == 0) {
     return false;
   }
@@ -57,7 +57,7 @@ bool cb_pop(circular_buffer_t* cb, uint8_t* value) {
   return true;
 }
 
-bool cb_peek(const circular_buffer_t* cb, uint8_t* value) {
+bool cb_peek(const circular_buffer_t* cb, uint32_t* value) {
   if (cb->size == 0 || !value) {
     return false;
   }
